encantamais-backend/requirements.txt
Django>=4.2
djangorestframework
gunicorn
whitenoise
psycopg2-binary
dj-database-url
python-dotenv
django-cors-headers
